requirements.txt
pandas
requests
python-dotenv
jinja2
sendgrid
openpyxl
